requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
deep-translator==1.11.4
openai>=1.12.0
googletrans>=3.1.0a0
tqdm>=4.66.1
langdetect>=1.0.9
pathlib>=1.0.1
typing>=3.7.4.3
asyncio>=3.4.3